+++ blog/content/posts/2024-10-08-tagging.md
@@ -5,7 +5,7 @@ image: https://images.unsplash.com/photo-1523309375637-b3f4f2347f2d?q=80&w=3732&
 description: We introduced the first Roq plugin, it is for collection tagging & with pagination support!
 author: ia3andy
 tags: plugin, frontmatter, guide, cool-stuff
-serie: roq-plugins
+series: roq-plugins
 ---
 
 My mind is getting blown by how much Quarkus was made for Static Site Generation. I just implemented a new plugin to generate tag pages and that was soooo easy.
