+++ blog/content/posts/2024-10-09-aliases.md
@@ -6,7 +6,7 @@ description: We introduced a way to declare aliases in FrontMatter. It is now ea
 author: mcruzdev
 tags: plugin, frontmatter, guide, cool-stuff
 aliases: [aliases-very-cool, aliases-4-ever, aliases-again]
-serie: roq-plugins
+series: roq-plugins
 ---
 
 In the last post, we saw how easy it is to use Quarkus for static site generator (@ia3andy's was right!). I am excited to share that we now have a new plugin that allows you to set up redirects for your blog posts! For this post, I've created three aliases.
