+++ blog/content/posts/2024-11-14-qrcode.md
@@ -6,7 +6,7 @@ description: Add a QR Code to your Roq website.
 author: ia3andy
 tags: qrcode, guide, cool-stuff
 date: 2024-11-14 14:00:00 +0200
-serie: roq-plugins
+series: roq-plugins
 ---
 
 Need to add a scannable QR Code to your website? Whether it's for a restaurant menu, event ticket, or any other use case where you want to make your content easily accessible via mobile devices, the Roq QR Code plugin has you covered.
