+++ blog/templates/layouts/roq-default/post.html
@@ -4,7 +4,7 @@ layout: theme-layouts/roq-default/post
 
 {@io.quarkiverse.roq.frontmatter.runtime.model.DocumentPage page}
 
-{#include partials/roq-serie /}
+{#include partials/roq-series /}
 
 {#insert /}
