--- plugin/serie/pom.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 https://maven.apache.org/xsd/maven-4.0.0.xsd">
-    <modelVersion>4.0.0</modelVersion>
-
-    <parent>
-        <groupId>io.quarkiverse.roq</groupId>
-        <artifactId>quarkus-roq-plugin-parent</artifactId>
-        <version>999-SNAPSHOT</version>
-        <relativePath>../pom.xml</relativePath>
-    </parent>
-    <artifactId>quarkus-roq-plugin-serie-parent</artifactId>
-    <packaging>pom</packaging>
-    <name>Quarkus Roq - Plugin - Serie</name>
-
-    <properties>
-        <quarkus-barcode.version>0.0.3</quarkus-barcode.version>
-    </properties>
-    <modules>
-        <module>deployment</module>
-        <module>runtime</module>
-    </modules>
-</project>
