--- plugin/serie/runtime/src/main/resources/META-INF/quarkus-extension.yaml
@@ -1,24 +1,0 @@
-name: Roq Plugin Series
-description: This plugin allows you to join multiple posts in a serie.
-metadata:
-  keywords:
-    - website
-    - ssg
-    - static
-    - generate
-    - jekyll
-    - hugo
-    - lume
-    - site
-    - generator
-    - web
-    - roq
-    - plugin
-    - serie
-  guide: https://docs.quarkiverse.io/quarkus-roq/dev/quarkus-roq-plugins.html#plugin-serie
-  icon-url: "https://raw.githubusercontent.com/quarkiverse/quarkus-roq/main/docs/modules/ROOT/assets/images/roq.svg"
-  categories:
-    - "web"
-  config:
-    - "quarkus.roq"
-  status: "preview"
