--- plugin/serie/runtime/src/main/resources/messages/serie_en.properties
@@ -1,3 +1,0 @@
-header=This post is part of the posts serie : {serieName}
-navigate_to=Navigate to
-select_post=Select a post
