--- plugin/serie/runtime/src/main/resources/messages/serie_fr.properties
@@ -1,3 +1,0 @@
-header=Ce poste fait partie de la série d'articles : {serieName}
-navigate_to=Aller à
-select_post=Sélectionner un article
