--- plugin/serie/runtime/src/main/resources/templates/partials/roq-serie.html
@@ -1,15 +1,0 @@
-{@io.quarkiverse.roq.frontmatter.runtime.model.Page page}
-
-{#if cdi:series}
-{#let serie=cdi:series.getForPage(page)}
-{#if serie}
-<section class="serie">
-  {serie:header(serie.title)} </br>
-  <label>{serie:navigate_to}</label>
-  {#for doc in serie.documents}
-  <a href="{doc.url}" {#if doc.id == page.id} class="active" {/if}>{doc.title}</a>
-  {/for}
-</section>
-{/if}
-{/let}
-{/if}
