+++ plugin/series/runtime/src/main/resources/messages/series_en.properties
@@ -1,0 +1,1 @@
+header={title} ({count} Parts Series)
