+++ plugin/series/runtime/src/main/resources/messages/series_fr.properties
@@ -1,0 +1,1 @@
+header={title} ({count} parties)
