+++ plugin/series/runtime/src/main/resources/templates/partials/roq-series.html
@@ -1,0 +1,26 @@
+{@io.quarkiverse.roq.frontmatter.runtime.model.Page page}
+
+{#if cdi:series}
+{#let series=cdi:series.get(page)}
+{#if series}
+<section class="series">
+  <h3>{series:header(series.title, series.documents.size)}</h3>
+  <ul>
+    {#for doc in series.documents}
+    <li>
+      {#if doc.id == page.id}
+      <span class="index">{doc_count}</span>
+      <span class="title">{doc.title}</span>
+      {#else}
+      <a href="{doc.url}">
+        <span class="index">{doc_count}</span>
+        <span class="title">{doc.title}</span>
+      </a>
+      {/if}
+    </li>
+    {/for}
+  </ul>
+</section>
+{/if}
+{/let}
+{/if}
