+++ theme/default/src/main/web/roq.scss
@@ -161,6 +161,9 @@ table tfoot td {
 /* - - - - - - - - - - Tag Styles - - - - - - - - - - */
 @import 'scss/parts/tag';
 
+/* - - - - - - - - - - Series Styles - - - - - - - - - - */
+@import 'scss/parts/series';
+
 /* - - - - - - - - - - Media Styles - - - - - - - - - - */
 @import 'scss/media';
