+++ theme/default/src/main/web/scss/parts/_series.scss
@@ -1,0 +1,31 @@
+.series {
+  li {
+    margin-bottom: 5px;
+  }
+  li, li a {
+    display: flex;
+  }
+
+  a {
+    text-decoration: none;
+    font-weight: bold;
+  }
+
+  .index {
+    background-color: $dark-gray;
+    color: white;
+    border-radius: 50%;
+    display: inline-block;
+    width: 30px;
+    height: 30px;
+    line-height: 30px;
+    vertical-align: center;
+    text-align: center;
+    font-weight: bold;
+    margin-right: 10px;
+  }
+
+  a .index {
+    background-color: $body-color;
+  }
+}
